Rename Proxy to ContentExtractor
